refactor model, service and validator to a folder for better readability and maintainability
